# Timer Initialization and Accuracy Fix